Refactor utilities
