+++ src/clowder/cli/reset.py
@@ -12,8 +12,7 @@ from typing import List, Optional
 import clowder.util.formatting as fmt
 from clowder.clowder_controller import CLOWDER_CONTROLLER
 from clowder.config import Config
-from clowder.util.clowder_utils import (
-    add_parser_arguments,
+from clowder.model.util import (
     filter_projects,
     validate_project_statuses
 )
@@ -24,6 +23,8 @@ from clowder.util.decorators import (
     valid_clowder_yaml_required
 )
 from clowder.util.parallel import reset_parallel
+
+from .util import add_parser_arguments
 
 
 def add_reset_parser(subparsers: argparse._SubParsersAction): # noqa
