+++ src/clowder/cli/start.py
@@ -10,17 +10,18 @@ import argparse
 import clowder.util.formatting as fmt
 from clowder.clowder_controller import CLOWDER_CONTROLLER
 from clowder.config import Config
+from clowder.model.util import (
+    filter_projects,
+    validate_project_statuses
+)
 from clowder.util.connectivity import network_connection_required
 from clowder.util.decorators import (
     print_clowder_name,
     print_clowder_repo_status,
     valid_clowder_yaml_required
 )
-from clowder.util.clowder_utils import (
-    add_parser_arguments,
-    filter_projects,
-    validate_project_statuses
-)
+
+from .util import add_parser_arguments
 
 
 def add_start_parser(subparsers: argparse._SubParsersAction) -> None: # noqa
