+++ src/clowder/cli/yaml.py
@@ -9,13 +9,14 @@ import argparse
 
 import clowder.util.formatting as fmt
 from clowder.clowder_controller import CLOWDER_CONTROLLER
-from clowder.util.clowder_utils import add_parser_arguments
 from clowder.util.decorators import (
     print_clowder_name,
     print_clowder_repo_status,
     valid_clowder_yaml_required
 )
 from clowder.util.yaml import print_yaml
+
+from .util import add_parser_arguments
 
 
 def add_yaml_parser(subparsers: argparse._SubParsersAction) -> None: # noqa
